move packetTrailer files
