--- src/e2ee/packetTrailer.test.ts
@@ -1,55 +1,0 @@
-import { describe, expect, it } from 'vitest';
-import { appendPacketTrailer, extractPacketTrailer } from './packetTrailer';
-
-describe('packetTrailer', () => {
-  it('extracts user timestamp and frame id from packet trailer', () => {
-    const payload = Uint8Array.from([1, 2, 3, 4]);
-    const trailer = appendPacketTrailer(payload, 1_744_249_600_123_456n, 42);
-    const extracted = extractPacketTrailer(trailer);
-
-    expect(Array.from(extracted.data)).toEqual(Array.from(payload));
-    expect(extracted.metadata).toEqual({
-      userTimestamp: 1_744_249_600_123_456n,
-      frameId: 42,
-    });
-  });
-
-  it('extracts timestamp-only trailer when frameId is 0', () => {
-    const payload = Uint8Array.from([1, 2, 3, 4]);
-    const trailer = appendPacketTrailer(payload, 1_744_249_600_123_456n, 0);
-    const extracted = extractPacketTrailer(trailer);
-
-    expect(Array.from(extracted.data)).toEqual(Array.from(payload));
-    expect(extracted.metadata).toEqual({
-      userTimestamp: 1_744_249_600_123_456n,
-      frameId: 0,
-    });
-  });
-
-  it('extracts frameId-only trailer when timestamp is 0', () => {
-    const payload = Uint8Array.from([1, 2, 3, 4]);
-    const trailer = appendPacketTrailer(payload, 0n, 42);
-    const extracted = extractPacketTrailer(trailer);
-
-    expect(Array.from(extracted.data)).toEqual(Array.from(payload));
-    expect(extracted.metadata).toEqual({
-      userTimestamp: 0n,
-      frameId: 42,
-    });
-  });
-
-  it('returns data unchanged when both timestamp and frameId are 0', () => {
-    const payload = Uint8Array.from([1, 2, 3, 4]);
-    const result = appendPacketTrailer(payload, 0n, 0);
-
-    expect(Array.from(result)).toEqual(Array.from(payload));
-  });
-
-  it('passes frames through when there is no valid trailer', () => {
-    const payload = Uint8Array.from([1, 2, 3, 4, 5]);
-    const extracted = extractPacketTrailer(payload);
-
-    expect(Array.from(extracted.data)).toEqual(Array.from(payload));
-    expect(extracted.metadata).toBeUndefined();
-  });
-});
