+++ src/index.ts
@@ -63,7 +63,7 @@ import {
 import { getBrowser } from './utils/browserParser';
 
 export { RpcError, type RpcInvocationData, type PerformRpcParams } from './room/rpc';
-export type { PacketTrailerMetadata } from './e2ee/packetTrailer';
+export type { PacketTrailerMetadata } from './packetTrailer/types';
 export {
   PacketTrailerManager,
   type PacketTrailerOptions,
